requirements.txt
# ==2.1.2
django
# ==3.8.2
djangorestframework
# ==2.0.4
django-angular
# ==1.2.5
psycopg2
# ==0.10.1
python-dotenv
# ==0.3
bitly_api
pytz
gunicorn
django-heroku
# ==4.1.2
whitenoise
# ==2.0.0
social-auth-app-django
# ==2.3.3
coreapi
